[test] Moved `FilterDiagsRAII` class to ROOTUnitTestSupport.h
